feat(personal-website): refactor tailwindcss integration and clean up unused utilities
