+++ apps/personal-website/package.json
@@ -27,8 +27,7 @@
     "i18next-browser-languagedetector": "^8.0.0",
     "i18next-resources-to-backend": "^1.2.1",
     "js-cookie": "^3.0.5",
-    "react": "^18.3.1",
-    "tailwindcss": "^4.0.0-alpha.30"
+    "react": "^18.3.1"
   },
   "devDependencies": {
     "@astrojs/check": "^0.9.4",
